Implemented assembly of H0, S
Minor changes to main.py were necessary to make it work. Not sure if it makes sense to do it this way, I essentially set all the diagonal elements of the matrix created there to 1 so that the assembling routine will visit the diagonal and put together H0 and S while assembling Hfull.